# flightstatus 2/open-api-docs.yaml
openapi: 3.0.1
info:
  title: Flight Status APIs
  description: APIs to create and fetch flight statuses from Elasticsearch
  version: "1.0"
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
  /setStatus:
    post:
      tags:
      - flight-status-controller
      operationId: insertNewFlightStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightEvent'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FlightEvent'
  /getStatus:
    get:
      tags:
      - flight-status-controller
      operationId: getFlightStatus
      parameters:
      - name: searchDTO
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/SearchDTO'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightEvent'
  /all:
    get:
      tags:
      - flight-status-controller
      operationId: findAll
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    FlightEvent:
      required:
      - eventName
      type: object
      properties:
        id:
          type: string
          format: uuid
        flightNo:
          type: string
        eventName:
          type: string
        eventTime:
          pattern: "^([0-9]{4})-([0-1][0-9])-([0-3][0-9]) ([0-1][0-9]|[2][0-3]):([0-5][0-9]):([0-5][0-9])$"
          type: string
    SearchDTO:
      type: object
      properties:
        flightNo:
          type: string
        flightDate:
          pattern: "^\\d{4}\\-(0[1-9]|1[012])\\-(0[1-9]|[12][0-9]|3[01])$"
          type: string
        flightdate:
          type: string
          writeOnly: true